+++ kong/plugin.go
@@ -4,19 +4,20 @@ package kong
 // Read https://docs.konghq.com/gateway/latest/admin-api/#plugin-object
 // +k8s:deepcopy-gen=true
 type Plugin struct {
-	CreatedAt    *int            `json:"created_at,omitempty" yaml:"created_at,omitempty"`
-	ID           *string         `json:"id,omitempty" yaml:"id,omitempty"`
-	Name         *string         `json:"name,omitempty" yaml:"name,omitempty"`
-	InstanceName *string         `json:"instance_name,omitempty" yaml:"instance_name,omitempty"`
-	Route        *Route          `json:"route,omitempty" yaml:"route,omitempty"`
-	Service      *Service        `json:"service,omitempty" yaml:"service,omitempty"`
-	Consumer     *Consumer       `json:"consumer,omitempty" yaml:"consumer,omitempty"`
-	Config       Configuration   `json:"config,omitempty" yaml:"config,omitempty"`
-	Enabled      *bool           `json:"enabled,omitempty" yaml:"enabled,omitempty"`
-	RunOn        *string         `json:"run_on,omitempty" yaml:"run_on,omitempty"`
-	Ordering     *PluginOrdering `json:"ordering,omitempty" yaml:"ordering,omitempty"`
-	Protocols    []*string       `json:"protocols,omitempty" yaml:"protocols,omitempty"`
-	Tags         []*string       `json:"tags,omitempty" yaml:"tags,omitempty"`
+	CreatedAt     *int            `json:"created_at,omitempty" yaml:"created_at,omitempty"`
+	ID            *string         `json:"id,omitempty" yaml:"id,omitempty"`
+	Name          *string         `json:"name,omitempty" yaml:"name,omitempty"`
+	InstanceName  *string         `json:"instance_name,omitempty" yaml:"instance_name,omitempty"`
+	Route         *Route          `json:"route,omitempty" yaml:"route,omitempty"`
+	Service       *Service        `json:"service,omitempty" yaml:"service,omitempty"`
+	Consumer      *Consumer       `json:"consumer,omitempty" yaml:"consumer,omitempty"`
+	ConsumerGroup *ConsumerGroup  `json:"consumer_group,omitempty" yaml:"consumer_group,omitempty"`
+	Config        Configuration   `json:"config,omitempty" yaml:"config,omitempty"`
+	Enabled       *bool           `json:"enabled,omitempty" yaml:"enabled,omitempty"`
+	RunOn         *string         `json:"run_on,omitempty" yaml:"run_on,omitempty"`
+	Ordering      *PluginOrdering `json:"ordering,omitempty" yaml:"ordering,omitempty"`
+	Protocols     []*string       `json:"protocols,omitempty" yaml:"protocols,omitempty"`
+	Tags          []*string       `json:"tags,omitempty" yaml:"tags,omitempty"`
 }
 
 // PluginOrdering contains before or after instructions for plugin execution order